Add backup of original STM32F407ZE from V4.6 board
Extracted via ST-Link v2 mini and STM32CubeProgrammer with settings:

* ST-Link configuration (on the right site):
  - Port SWD
  - Frequency: 4000 kHz
  - Mode: Under reset
  - Access port; 0
  - Reset mode: Hardware reset
  - Speed: Reliable
  - Shared: Disabled
* Device memory (in the main view)
  - Address: 0x08000000
  - Size: 0x80000

All power was removed from the Qidi X-Max (even from the RPi). Following
Pins of the ST-Link v2 min was connected to the Board:

* 3.3V  -> 3V
* GND   -> SWI7
* SWDIO -> GND
* SWCLK -> SWC9

The red LED (D1) should get enabled when the ST-Link is connected to the
host PC (running STm32CubeProgrammer). The RST must be pressed (don't let
it go yet), the Connect button in STM32CubeProgrammer must be pressed and
then the RST button must be released immediately.

The green "Read" button can be pressed to read out the memory and the arrow
next to it can be used to save the content to a file.

Signed-off-by: Sven Eckelmann <sven@narfation.org>