bltouch: Work around timing problems between MCU and rpi
It seems like sometimes the RPi is still forcing the pin_down (or at least
the BLTouch is still doing this) when the Z stepper goes up. This caused
random errors during probing like:

  No trigger on probe after full movement

To work around this, just add more delay between probing and movement of
the bed.

It might be better to just try to use a spare Pin of the MCU. Maybe the
wifi module can be removed and its pin then be reused.

Signed-off-by: Sven Eckelmann <sven@narfation.org>