bltouch: Use wifi module pins for BLTouch control
The multi MCU (via rpi) klipper functionality wasn't working to well with
BLtouch because the pin_down signal was still sent when the bed moved. As a
workaround, it was possible to have the bed move further until it hit the
BLTouch.

But since the ESP8266EX based wifi module is not used anymore with Klipper,
just remove it and use one of its two serial pins as control pin:

                       WiFi
  +------+------+------+------+
  | VDD  | PG1  | VDD  | Vcap |
  +------+------+------+------+
  | PB10 | GND  | NC?  | GND  |
  +------+------+------+------+
  |                           |
  |                           |
  |                           |
  |    ESP8266EX              |
  |                           |
  |                           |
  |                           |
  +---------------------------+

Signed-off-by: Sven Eckelmann <sven@narfation.org>