Allow to power up board remotely
The "power-off-module" from qidi is directly switching the 230VAC and is
using the front-button to initially power up the microcontroller.

The button is connected to the three-port screw terminanal on the module.
From top to bottom (assuming that the signal + 5V connectors are on the
bottom):

* Common (green wire)
* NormalOpen (red wire)
* NormalClosed (blue wire)

It is possible to add a 250VAC rated relay module (KY-019 or similar) in
parallel. Just connect in Common to Common and NormalOpen to NormalOpen.

The 5V+GND need to be connected to the RaspberryPi. For the signal, GPIO26
was chosen because it is pulled-down by default.

Since the front button is a momentary switch, the simulated button press
will be released after 1s.

This modification requires working on parts with mains voltages and should
only be done by professionals. There is:

* Risk of deadly electical shock
* Risk of irreversable damage to other components attached

Signed-off-by: Sven Eckelmann <sven@narfation.org>