Handle overshooting of heater_bed
During the initial heating, the bed seems seems to cause some overshooting.
This can be damped slightly by increasing the smoothing time. But still,
slight overshoot still happens. But since a degree more during the initial
heatup seems to be acceptable (because the actual temperature will be hit
during the first layer), it is not necessary to enforce the exact
temperature in the startup macro.

Signed-off-by: Sven Eckelmann <sven@narfation.org>